Move geometric.py our of primitives and into core where it belongs.  Also update test_import_primitives.py
