+++ kcwidrp/core/geometric.py
